管理页面完善

+++ common/src/main/java/com/daoli/office/vo/sheng/tai/ShengtaiExamVo.java
@@ -26,7 +26,7 @@ public class ShengtaiExamVo {
 
     private String examDesc;
 
-    private Integer examScore;
+    private Float examScore;
 
     private String examStatus;
